Updated support for the Simple-Copy-Command
* updated ns-struct to TP 4065 2/20/2020
* fixed invalid use of zero-based value in pp
* Adjusted to support extended-lba
* Added test exclusively for the Simple-Copy-Command
* tests: added note on expansion of tests [scc]

Additionally, the NVMe spec. only defines a single copy-format, however,
based on the different means of transporting said format, then it is
advantageous to support e.g. simpler formats such as as an array of
source-LBAs, an array of (source-LBAs, nlb) tuples and many other
possible formats.

This can be supported by different means, e.g. format-conversion and
other things, in case a given format is not supported by the device, it
will incur a cost, but it will work.

To strive in that direction, then commit will be redo with a different
``lblk_cmd_scopy`` signature.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>